fix: replace ssrSafeFirstChildSelector with ssrSafeNotFirstChildSelector
BREAKING CHANGE: The ssrSafeFirstChildSelector actually never worked 
correctly. There isn't a fix for this so it has been removed. It has 
been replaced by the ssrSafeNotFirstChildSelector which does the 
opposite. It will apply if the element is not the first child of its 
parent. It can be used in the same situation as the previous 
implementation by switching the around the default style and the 
override since they now apply to the other set of elements.