feat: responsive sizing for Dialogs
BREAKING CHANGE: The way the width of the Dialog component has changed. 
Width is no longer an optional string, and instead is a mandatory number 
used to set the width in pixels. Check the documentation to what the 
behaviour of this prop is. The Dialog component will now shrink to fit 
the screen if necessary. This could affect the layout or positioning 
inside the Dialog, however if the included Dialog.Header, Dialog.Body, 
and Dialog.Footer components are already being used, the impact should 
be minimal.